container/list: move List[T] type out of x/tls/x509utils/certpool
