+++ tls/go.mod
@@ -8,6 +8,7 @@ require (
 )
 
 require (
+	darvaza.org/x/container v0.0.0-00010101000000-000000000000
 	github.com/zeebo/blake3 v0.2.4
 	golang.org/x/crypto v0.29.0
 )
@@ -18,3 +19,5 @@ require (
 	golang.org/x/sys v0.27.0 // indirect
 	golang.org/x/text v0.20.0 // indirect
 )
+
+replace darvaza.org/x/container => ../container
