+++ tls/x509utils/certpool/utils.go
@@ -7,6 +7,7 @@ import (
 	"io/fs"
 
 	"darvaza.org/core"
+	"darvaza.org/x/container/list"
 	"darvaza.org/x/tls/x509utils"
 )
 
@@ -35,12 +36,12 @@ func copyMap[K comparable, V any](m map[K]V, fn func(V) (V, bool)) map[K]V {
 	return out
 }
 
-func copyMapList[K comparable, V any](m map[K]*List[V], fn func(V) (V, bool)) map[K]*List[V] {
+func copyMapList[K comparable, V any](m map[K]*list.List[V], fn func(V) (V, bool)) map[K]*list.List[V] {
 	if fn == nil {
 		fn = copyFn
 	}
 
-	out := make(map[K]*List[V], len(m))
+	out := make(map[K]*list.List[V], len(m))
 	for k, l1 := range m {
 		if l2 := l1.Copy(fn); l2.Len() > 0 {
 			out[k] = l2
@@ -66,16 +67,16 @@ func sliceForEach[T any](ctx context.Context, fn func(context.Context, T) bool, 
 	}
 }
 
-func appendMapList[K, V comparable](m map[K]*List[V], key K, value V) {
+func appendMapList[K, V comparable](m map[K]*list.List[V], key K, value V) {
 	l, ok := m[key]
 	if !ok {
-		l = new(List[V])
+		l = new(list.List[V])
 		m[key] = l
 	}
 	l.PushBack(value)
 }
 
-func deleteMapListMatchFn[K, V comparable](m map[K]*List[V], keys []K, eq func(v V) bool) {
+func deleteMapListMatchFn[K, V comparable](m map[K]*list.List[V], keys []K, eq func(v V) bool) {
 	for _, key := range keys {
 		if l, ok := m[key]; ok {
 			l.DeleteMatchFn(eq)
